move collection name header to main content area
Display collection name as a page header at the top of the docs content
area with the Fetch in Bruno button beside it, instead of in the sidebar.
Drop the redundant markdown H1 from the sample collection now that the
name renders as its own header.
